Set number input step to "any"
The step value for the x and y range were not set, which is why firefox
would highlight the fields in red and issue the warning to "Please
select a valid value" when a float was entered in any of the fields. I
believe this is an html5 standard, but not all browsers enforce it.