Remove "Upload plot data" button.
The "Upload plot data" button allowed a user to replace one data set
with another. Although it might be somewhat useful to effectively remove
and add a data set with one button press instead of two, the potential
for confusion seems to outweigh the benefit, especially if a user
accidentally selects the wrong data set (e.g. replace A.dat with B.dat).