Discover react data path from tempfile.
This commit enables browser_plot to always pull in the latest changes
made to CellBlender. Previously, the list of available files could be
wrong. You could potentially have either files present in the reaction
directory but missing from the list or files in the list but missing
from the reaction directory.

We now can get around this problem by passing the location of a
temporary file to the server when it first starts. This file will
contain the location of the reaction directory. If the location ever
changes, we can update it from CellBlender. So both browser_plot and CB
know this location for the duration that they run (even if you make
changes to the existing blend file or open different blend files), and
it acts as a way for CB to communicate with the server.

Also, the seed value is used as part of a data set's name. This is
particularly useful when running multiple seeds.

Note: this change depends on a commit from CB which hasn't been made
yet.